pyproject.toml
[project]
name = "ai-engineer-roadmap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.72.0",
    "fastapi>=0.128.0",
    "langchain>=1.2.9",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-openai>=1.1.7",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.0",
    "lark>=1.3.1",
    "llama-index>=0.14.13",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-llms-openai>=0.6.18",
    "llama-index-vector-stores-qdrant>=0.9.1",
    "matplotlib>=3.10.8",
    "openai>=2.16.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.2",
    "tiktoken>=0.12.0",
]
